+++ scalingomock/databasesservice_mock.go
@@ -65,6 +65,22 @@ func (mr *MockDatabasesServiceMockRecorder) DatabaseEnableFeature(arg0, arg1, ar
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DatabaseEnableFeature", reflect.TypeOf((*MockDatabasesService)(nil).DatabaseEnableFeature), arg0, arg1, arg2, arg3)
 }
 
+// DatabaseListMaintenance mocks base method.
+func (m *MockDatabasesService) DatabaseListMaintenance(arg0 context.Context, arg1, arg2 string, arg3 scalingo.PaginationOpts) ([]*scalingo.Maintenance, scalingo.PaginationMeta, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "DatabaseListMaintenance", arg0, arg1, arg2, arg3)
+	ret0, _ := ret[0].([]*scalingo.Maintenance)
+	ret1, _ := ret[1].(scalingo.PaginationMeta)
+	ret2, _ := ret[2].(error)
+	return ret0, ret1, ret2
+}
+
+// DatabaseListMaintenance indicates an expected call of DatabaseListMaintenance.
+func (mr *MockDatabasesServiceMockRecorder) DatabaseListMaintenance(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DatabaseListMaintenance", reflect.TypeOf((*MockDatabasesService)(nil).DatabaseListMaintenance), arg0, arg1, arg2, arg3)
+}
+
 // DatabaseShow mocks base method.
 func (m *MockDatabasesService) DatabaseShow(arg0 context.Context, arg1, arg2 string) (scalingo.Database, error) {
 	m.ctrl.T.Helper()
@@ -78,6 +94,36 @@ func (m *MockDatabasesService) DatabaseShow(arg0 context.Context, arg1, arg2 str
 func (mr *MockDatabasesServiceMockRecorder) DatabaseShow(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DatabaseShow", reflect.TypeOf((*MockDatabasesService)(nil).DatabaseShow), arg0, arg1, arg2)
+}
+
+// DatabaseShowMaintenance mocks base method.
+func (m *MockDatabasesService) DatabaseShowMaintenance(arg0 context.Context, arg1, arg2, arg3 string) (scalingo.Maintenance, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "DatabaseShowMaintenance", arg0, arg1, arg2, arg3)
+	ret0, _ := ret[0].(scalingo.Maintenance)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// DatabaseShowMaintenance indicates an expected call of DatabaseShowMaintenance.
+func (mr *MockDatabasesServiceMockRecorder) DatabaseShowMaintenance(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DatabaseShowMaintenance", reflect.TypeOf((*MockDatabasesService)(nil).DatabaseShowMaintenance), arg0, arg1, arg2, arg3)
+}
+
+// DatabaseUpdateMaintenanceWindow mocks base method.
+func (m *MockDatabasesService) DatabaseUpdateMaintenanceWindow(arg0 context.Context, arg1, arg2 string, arg3 scalingo.MaintenanceWindowParams) (scalingo.Database, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "DatabaseUpdateMaintenanceWindow", arg0, arg1, arg2, arg3)
+	ret0, _ := ret[0].(scalingo.Database)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// DatabaseUpdateMaintenanceWindow indicates an expected call of DatabaseUpdateMaintenanceWindow.
+func (mr *MockDatabasesServiceMockRecorder) DatabaseUpdateMaintenanceWindow(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DatabaseUpdateMaintenanceWindow", reflect.TypeOf((*MockDatabasesService)(nil).DatabaseUpdateMaintenanceWindow), arg0, arg1, arg2, arg3)
 }
 
 // DatabaseUpdatePeriodicBackupsConfig mocks base method.